requirements.txt
altair==5.2.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
flatbuffers==23.5.26
gitdb==4.0.11
GitPython==3.1.41
humanfriendly==10.0
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.4
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
onnx==1.15.0
onnxruntime==1.16.3
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.0
patchify==0.2.3
pillow==10.2.0
protobuf==4.25.2
pyarrow==14.0.2
pydeck==0.8.1b0
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.32.1
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.11.4
six==1.16.0
smmap==5.0.1
streamlit==1.30.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tifffile==2023.12.9
toml==0.10.2
toolz==0.12.0
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
